requirements.txt
absl-py==2.2.2
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
braceexpand==0.1.7
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
confection==0.1.5
contexttimer==0.3.3
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
datasets==3.6.0
decorator==5.2.1
decord==0.6.0
dill==0.3.8
distlib==0.3.9
einops==0.8.1
et_xmlfile==2.0.0
evaluate==0.4.0
exceptiongroup==1.3.0
executing==2.2.0
fairscale==0.4.4
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
hf-xet==1.1.0
huggingface-hub==0.31.1
identify==2.6.10
idna==3.10
imageio==2.37.0
iopath==0.1.10
ipython==8.36.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kaggle==1.7.4.5
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
murmurhash==1.0.13
narwhals==1.40.0
networkx==3.4.2
nltk==3.6.6
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
openai==0.23.0
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
openpyxl==3.1.5
packaging==24.2
pandas==1.4.3
pandas-stubs==2.0.1.230501
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.1
portalocker==3.1.1
pre_commit==4.2.0
preshed==3.0.9
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycocoevalcap==1.2
pycocotools==2.0.8
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-magic==0.4.27
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
responses==0.18.0
rich==14.0.0
rouge==1.0.1
rouge_score==0.1.2
rpds-py==0.25.1
safetensors==0.5.3
salesforce-lavis==1.0.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==2.2.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
spacy==3.8.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
streamlit==1.45.1
sympy==1.14.0
tenacity==9.1.2
text-unidecode==1.3
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==0.4.12
tokenizers==0.13.3
toml==0.10.2
torch==2.7.0
torchvision==0.22.0
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
transformers==4.26.1
triton==3.3.0
typer==0.15.4
types-pytz==2025.2.0.20250326
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
virtualenv==20.31.2
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webdataset==0.2.111
webencodings==0.5.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
